Adaptation of IssuesList, IssuesListItem added
