+++ plugins/tracker-assets/lang/ru.json
@@ -87,7 +87,7 @@
     "Space": "",
     "SetDueDate": "Указать срок выполнения\u2026",
     "ChangeDueDate": "Изменить срок выполнения\u2026",
-    "ModificationDate": "Изменино {value}",
+    "ModificationDate": "Изменено {value}",
     "Team": "",
     "Issue": "Задача",
     "Document": "",
